requirements.txt
pip~=25.1.1
Django==5.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
channels==4.0.0
channels-redis==4.1.0
redis==4.6.0
daphne==4.0.0